65.Creates user's seeder and deletes user frontend side

+++ src/app/users/users.component.ts
@@ -23,14 +23,18 @@ export class UsersComponent implements OnInit {
   }
 
   onDeleteUser(user: User) {
-
-    this.userService.deleteUser(user).subscribe(
-      response => {
-
+    const deleteUser = confirm('Do you really want to delete user '+
+      user.name + ' ' + user.lastname + '');
+    if(deleteUser) {
+      this.userService.deleteUser(user).subscribe(
+        response => {
+        const idx = this.users.indexOf(user);
+        this.users.splice(idx, 1);
           alert(response['message']);
 
-      }
-    )
+        }
+      );
+    }
   }
 
   onSelectUser(user: User) {
